moved auto canni to Classes/Shaman.cs EXP Alert updates for live making sure all threads are in EN-US to handle numbers in non us countries
